megaphone reboost

+++ ex00/megaphone.cpp
@@ -6,7 +6,7 @@
 /*   By: achansar <achansar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/13 15:19:37 by achansar          #+#    #+#             */
-/*   Updated: 2023/07/28 13:36:05 by achansar         ###   ########.fr       */
+/*   Updated: 2023/07/28 13:41:17 by achansar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ int main(int argc, char **argv)
     if (argc == 1)
         std::cout << "* LOUD AND UNBEARABLE FEEDBACK NOISE *" << std::endl;
     else {
-        for (size_t j = 1; j < argc; j++) {
+        for (int j = 1; j < argc; j++) {
             for (size_t i = 0; i < strlen(argv[j]); i++)
                 std::cout << (char)toupper(argv[j][i]);
         }
